// src/types/MetaData.ts
export interface MetaData {
    og: {
        title: string;
        description: string;
        url: string;
        image: string;
        type: string;
    };
    twitter: {
        card: string;
        domain: string;
        title: string;
        description: string;
        url: string;
        image: string;
    };
}